+++ web/src/engine/Crypto_test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { GetHexFromBytes, GetBytesFromHex } from './BufferEncoder';
 import * as testee from './Crypto';
+import { MD5 } from './Crypto';
 
 // TODO: ⚠️ Use constant expected values instead of using business logic to calculate expected values
 
@@ -14,6 +15,12 @@ describe('Hashing', () => {
     it('Should properly hash a string using SHA-512', async () => {
         const hash = await testee.HashUTF8('SHA-512', 'abc');
         expect(GetHexFromBytes(hash)).toBe('ddaf35a193617abacc417349ae204131' + '12e6fa4e89a97ea20a9eeee64b55d39a' + '2192992a274fc1a836ba3c23a3feebbd' + '454d4423643ce80e2a9ac94fa54ca49f');
+    });
+
+    it('Should correctly compute MD5 hash of a Uint8Array', () => {
+        const hash = MD5(new Uint8Array([104, 101, 108, 108, 111]));
+        const expectedHex = '5d41402abc4b2a76b9719d911017c592';
+        expect(GetHexFromBytes(hash)).toBe(expectedHex);
     });
 });
