+++ web/src/engine/websites/iqiyi_e2e.ts
@@ -6,13 +6,13 @@ new TestFixture({
         title: 'iqiyi'
     },
     container: {
-        url: 'https://www.iqiyi.com/manhua/detail_18yzmrvu2t.html',
-        id: '/manhua/detail_18yzmrvu2t.html',
+        url: 'https://manhua.iqiyi.com/comic/reader?comicId=243860070&episodeId=&from=',
+        id: '243860070',
         title: '苍穹榜之圣灵纪',
     },
     child: {
-        id: '/manhua/reader/18yzmrvu2t_18yzho1c7d.html',
-        title: '1 第1话：灵路榜（上）',
+        id: '1007800170',
+        title: '第1话：灵路榜（上）',
     },
     entry: {
         index: 2,
